refactor: merge LottoNumber class Number class
