+++ README.md
@@ -18,6 +18,7 @@
     * 숫자로만 이루어져 있는지 확인
     * 1이상 45이하인지 확인
 * 지난 주 당첨 번호 입력받기
+    * 당첨 번호 클래스(WinningLotto)는 싱글톤 클래스
     * 1이상 45이하의 자연수만 받기
     * 쉼표로 당첨 번호 구분
         * 1, 2, 3, 4, 5, 6 -> [1, 2, 3, 4, 5, 6]
